+++ collectors/bzimport/convertors.py
@@ -613,6 +613,7 @@ class FlawConvertor(BugzillaGroupsConvertorMixin):
         meta_attr["alias"] = self.alias
         meta_attr["depends_on"] = self.depends_on
         meta_attr["related_cves"] = [c for c in self.cve_ids if c != cve_id]
+        meta_attr["bz_summary"] = self.flaw_bug["summary"]
         meta_attr["last_change_time"] = self.flaw_bug["last_change_time"]
         meta_attr["last_imported_dt"] = timezone.now()
         meta_attr["acl_labels"] = self.groups