+++ docs/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Erratum advisory name to flaw filter (OSIDB-922)
 - CORS allow-list functionality (OSIDB-967, OSIDB-965)
 
+### Changed
+- Set Jira trackers as public instead of embargoed when private (OSIDB-1013)
+
 ## [3.2.2] - 2023-06-19
 ### Changed
 - Account for TRIAGE in the title/summary (OSIDB-999)