+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement adding new flaw comments (OSIDB-81)
 - Erratum advisory name to flaw filter (OSIDB-922)
 - CORS allow-list functionality (OSIDB-967, OSIDB-965)
+- Raw bugzilla summary to Flaw.meta_attr (OSIDB-1016)
 
 ### Changed
 - Set Jira trackers as public instead of embargoed when private (OSIDB-1013)